MinorTweak
So it's still crashing but I know why I believe. Firstly I fixed the calculation for getting the mouse position in world space because drawline showed me that was extremely wrong but it's clear to me the issue is how I am doing the AI the tutorials I looked at must be wrong or I must of missed something important because.

It's the AI module that is crashing, direction I believe is correct now

What I think is happening we are calling the AIController I made and it's empty. According to what I saw it was saying this isn't wrong but it's always felt wrong. I need to look into the AI part of this more and possibly redo it completely. There is something I am not getting.